# Resound-UI/Builtiful-Interface/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  description: Resound customer signal intelligence API
  version: 0.1.0
paths:
...
  /signals:
    get:
      tags:
      - signals
      summary: List Signals
      operationId: listSignals
      parameters:
      - name: brandId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Brandid
      - name: source
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
      - name: area
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Area
      - name: severity
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
      - name: sentiment
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sentiment
      - name: period
        in: query
        required: false
        schema:
          enum:
          - 24h
          - 7d
          - 30d
          - qtd
          type: string
          default: 7d
          title: Period
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      - name: X-Resound-Organization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Organization
      - name: X-Resound-Team
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Team
      - name: X-Resound-User
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /signals/critical:
    get:
      tags:
      - signals
      summary: Get Critical Signals
      operationId: getCriticalSignals
      parameters:
      - name: brandId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Brandid
      - name: period
        in: query
        required: false
        schema:
          enum:
          - 24h
          - 7d
          - 30d
          - qtd
          type: string
          default: 7d
          title: Period
      - name: X-Resound-Organization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Organization
      - name: X-Resound-Team
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Team
      - name: X-Resound-User
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SignalDetail'
                title: Response Getcriticalsignals
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /signals/{signalId}:
    get:
      tags:
      - signals
      summary: Get Signal
      operationId: getSignal
      parameters:
      - name: signalId
        in: path
        required: true
        schema:
          type: integer
          title: Signalid
      - name: X-Resound-Organization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Organization
      - name: X-Resound-Team
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Team
      - name: X-Resound-User
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /workflows/source-sync:
    post:
      tags:
      - workflows
      summary: Start Source Sync
      operationId: startSourceSync
      parameters:
      - name: X-Resound-Organization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Organization
      - name: X-Resound-Team
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-Team
      - name: X-Resound-User
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Resound-User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceSyncInput'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowJob'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Classification:
      properties:
        id:
          type: integer
          title: Id
        signalId:
          type: integer
          title: Signalid
        isAboutBrand:
          type: boolean
          title: Isaboutbrand
        area:
          type: string
          title: Area
        subarea:
          anyOf:
          - type: string
          - type: 'null'
          title: Subarea
        sentiment:
          type: string
          enum:
          - negative
          - neutral
          - positive
          - mixed
          title: Sentiment
        severity:
          type: string
          enum:
          - low
          - medium
          - high
          - critical
          title: Severity
        actionClass:
          type: string
          enum:
          - immediate
          - sprint
          - roadmap
          - fyi
          - ignore
          title: Actionclass
        rootCauseHypothesis:
          type: string
          title: Rootcausehypothesis
        summary:
          type: string
          title: Summary
        confidence:
          type: number
          title: Confidence
      type: object
      required:
      - id
      - signalId
      - isAboutBrand
      - area
      - sentiment
      - severity
      - actionClass
      - rootCauseHypothesis
      - summary
      - confidence
      title: Classification
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Route:
      properties:
        id:
          type: integer
          title: Id
        signalId:
          type: integer
          title: Signalid
        classificationId:
          type: integer
          title: Classificationid
        owner:
          type: string
          title: Owner
        ruleMatched:
          anyOf:
          - type: string
          - type: 'null'
          title: Rulematched
        confidence:
          type: number
          title: Confidence
        reroutedFrom:
          anyOf:
          - type: string
          - type: 'null'
          title: Reroutedfrom
        createdAt:
          type: string
          title: Createdat
      type: object
      required:
      - id
      - signalId
      - classificationId
      - owner
      - ruleMatched
      - confidence
      - createdAt
      title: Route
...
    Signal:
      properties:
        id:
          type: integer
          title: Id
        brandId:
          type: string
          title: Brandid
        source:
          type: string
          title: Source
        externalId:
          type: string
          title: Externalid
        url:
          type: string
          title: Url
        authorHandle:
          type: string
          title: Authorhandle
        authorMeta:
          anyOf:
          - type: string
          - type: 'null'
          title: Authormeta
        reach:
          anyOf:
          - type: integer
          - type: 'null'
          title: Reach
        content:
          type: string
          title: Content
        postedAt:
          type: string
          title: Postedat
        createdAt:
          type: string
          title: Createdat
      type: object
      required:
      - id
      - brandId
      - source
      - externalId
      - url
      - authorHandle
      - content
      - postedAt
      - createdAt
      title: Signal
    SignalDetail:
      properties:
        signal:
          $ref: '#/components/schemas/Signal'
        classification:
          $ref: '#/components/schemas/Classification'
        route:
          $ref: '#/components/schemas/Route'
        patternId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Patternid
        patternName:
          anyOf:
          - type: string
          - type: 'null'
          title: Patternname
      type: object
      required:
      - signal
      - classification
      - route
      title: SignalDetail
    SignalList:
      properties:
        signals:
          items:
            $ref: '#/components/schemas/SignalDetail'
          type: array
          title: Signals
        total:
          type: integer
          title: Total
      type: object
      required:
      - signals
      - total
      title: SignalList
...
    SourceSyncInput:
      properties:
        brandId:
          type: string
          title: Brandid
      type: object
      required:
      - brandId
      title: SourceSyncInput
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WorkflowJob:
      properties:
        id:
          type: integer
          title: Id
        workflowId:
          type: string
          title: Workflowid
        runId:
          anyOf:
          - type: string
          - type: 'null'
          title: Runid
        workflowType:
          type: string
          title: Workflowtype
        status:
          type: string
          title: Status
        taskQueue:
          anyOf:
          - type: string
          - type: 'null'
          title: Taskqueue
        createdAt:
          type: string
          title: Createdat
      type: object
      required:
      - id
      - workflowId
      - workflowType
      - status
      - createdAt
      title: WorkflowJob
servers:
- url: /api
  description: Base API path